[TASK] Refactor commandline option default values
Remove hasArgument() checks on setting Import object values:
It always returns true - even if the argument or option
has not been passed on the commandline.
Remove option checks on setting Export object values:
getArgument() and getOption() return false / null / [] or
the default value if not passed on commandline. The nulls
have to be casted to match the strict typing of the receiving
object fields.
Set option default values by Import and Export object
default values.
In functional tests object default values have not been available in
ExportCommand::configure() yet but only in ExportCommand::execute()
with the current stubbing method (interesting ..), thus it was
replaced by the GeneralUtility::addInstance() stubbing approach.