[BUGFIX] Handle "Undefined index" in exportAddRecordsFromRelations()
If there is no record available, then member variable $this->dat is
an empty array and property $this->dat['records'] does not exist.
If there is at least one record, the property $this->dat['records']
is always an array: No need to check $this->dat['records'] for
type array though.