[TASK] Refactor ImportExport::doesRecordExist()
Rename to getRecordFromDatabase() to be more
precise regarding its functionality.
Rename $recInf to $recordDb to align with local variable
$record.
Rename $doesRE to $recordDb to match the common
wording.
Initialize $fields in method signature instead of
method body as is the normal use case.
Initialize $fields in the method signature instead
of the method body, as is the normal use case.
Use strict typing when using $recordDb in conditions.
Align some comments.