[TASK] Refactor ImportExport::addFiles()
Rename $fileInfo to $fileStat to make $fileInfo available
for different context.
Rename $fileRecord to $fileInfo where it is not about a
database record of sys_file* but about information of the
file itself.
Pass the relations array by reference instead of using a copy
to reduce memory consumption.
Align some comments.