[TASK] Refactor ImportExport::renderSelectBox()
Rename $prefix to $name to match the common wording.
Rename $optValues to $options to be shorter and
cleaner.
Pass the options array by reference instead of
using a copy to reduce memory consumption.
Rename $opt to $optionsHtml to align with similar
variables.
Rename $isSelFlag to $isValueInOptions to be more
verbose.
Remove cast of $value as the variable is of strict
type string already.
Rename $sel to $selectedHtml to be more verbose and
in line with $optionsHtml.
Use sprintf() instead of string concatenation to align with
similar variables.
Align some comments.