[TASK] Refactor ImportExport::compareRecords()
Moved the initialization of the diff utility object
to a new subroutine getDiffUtility().
Pass the $databaseRecord by reference instead of
using a copy to reduce memory consumption.
Rename $inverseDiff to $inverse as it is shorter and
sufficient.
Remove duplicate rendering of html output of $output
and $tRows and replace by single $diffHtml.
Rename $fN to $fieldName to match the common wording.
Rename $value to $fieldValue to match the common wording.
Align some comments.