[BUGFIX] Initialize storages before call of checkImportPrerequisites()
The check no.2 of checkImportPrerequisites() could not yet
have worked because the member variable $this->storages has not
yet been initialised. This has been fixed by initializing it
before performing the check.

This has a side effect: A storage record is created in the
database if no local default storage already exists -
e.g. during tests.