[BUGFIX] Fix callback processSoftReferencesFlexFormCallBack()
The parameters $dataValue_ext1 and $dataValue_ext2 can be null
and thus cannot have applied a type declaration. This is a
regression which has been introduced with the strict typing of
this class.
The method gets called with a seventh parameter
$workspaceOptions.
The variable $origUid can be of type integer and thus gets
casted to string in order to match the strict typing of
processSoftReferencesSubstTokens().