[BUGFIX] Consider new record UIDs when sorting records on update
When importing records via update, the relations between imported records
and imported pages have formerly been set using the record UIDs of the
import file, but these UIDs have probably been changed during import and
thus the new UIDs have to be used. This is in line with sorting pages in
writePagesOrder().