[TASK] Remove temporarily required string cast of record ID
This cast has been introduced to satisfy the strict type of the parameter
of strpos() which has been used in writePagesOrder() to check the record ID
for "NEW". strpos() has been replaced in the meantime by
MathUtility::canBeInterpretedAsInteger() which can handle both strings and
integers and thus the cast can be removed again.