[BUGFIX] Cleanup former exports properly
A repeating export with the same name and saving the related
files in an external folder has failed as the export routine
wanted to create an already existing external folder.
In order to fix it and to have never any remains of former
exports, the external folder and its files get removed before
a new export file gets created.